refactor: extract filter contacts logic to hook
